Extracted wrap and withdraw, to make them testable
