--- apps/eth-wrapper/src/typings/fonts.d.ts
@@ -1,2 +1,0 @@
-declare module '*.woff';
-declare module '*.woff2';
